+++ node/chainctx/beacon_chain.go
@@ -9,6 +9,7 @@ package chainctx
 import (
 	"gitlab.com/jaxnet/jaxnetd/node/blocknodes"
 	"gitlab.com/jaxnet/jaxnetd/types/chaincfg"
+	"gitlab.com/jaxnet/jaxnetd/types/pow"
 	"gitlab.com/jaxnet/jaxnetd/types/wire"
 )
 
@@ -33,5 +34,6 @@ func (c *beaconChain) GenesisBlock() *wire.MsgBlock { return c.chainParams.Genes
 func (c *beaconChain) GenesisBeaconHeight() int32   { return 0 }
 
 func (c *beaconChain) NewNode(blockHeader wire.BlockHeader, parent blocknodes.IBlockNode, serialID int64) blocknodes.IBlockNode {
-	return blocknodes.NewBeaconBlockNode(blockHeader, parent, serialID)
+	powWeight := pow.CalcPowWeight(c.chainParams.PowParams.PowLimit, blockHeader.Bits(), c.chainParams.PowParams.HashSortingSlotNumber)
+	return blocknodes.NewBeaconBlockNode(blockHeader, parent, serialID, powWeight)
 }